pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
fal-client = "*"
requests = "*"
google-cloud-texttospeech = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

